refactor: project 모델 타입 리팩토링
