+++ src/api/index.ts
@@ -1,5 +1,5 @@
 //user
-export type { User } from '@/api/services/user/model'
+export type { User, UserResponse } from '@/api/services/user/model'
 export { stackNames } from '@/api/services/user/model'
 
 //====================
