+++ src/api/services/project/index.ts
@@ -4,7 +4,7 @@ import {
   usePutMutation,
   useDeleteMutation,
 } from '@/api/lib/fetcher'
-import { Project } from '@/api/mocks/handlers/project'
+import { Project } from '@/api/services/project/model'
 
 // 프로젝트 생성 응답 타입 정의
 export interface ProjectCreateResponse {
