+++ src/api/services/question/types.ts
@@ -3,7 +3,7 @@ import {
   CSQuestion,
   HistoryByDate,
   QuestionHistoryItem,
-} from '@/api/mocks/handlers/project'
+} from '@/api/services/project/model'
 
 // CS 질문 생성 요청 타입 정의
 export interface CreateProjectCsQuestionRequest {
